test: deduplication
