+++ src/rulesets/oas/__tests__/oas3-valid-schema-example.ts
@@ -20,11 +20,11 @@ describe('oas3-valid-schema-example', () => {
     });
   });
 
-  describe('components', () => {
+  describe.each(['components', 'headers'])('%s', field => {
     test('will pass when simple example is valid', async () => {
       const results = await s.run({
         openapi: '3.0.2',
-        components: {
+        [field]: {
           schemas: {
             xoxo: {
               type: 'string',
@@ -39,7 +39,7 @@ describe('oas3-valid-schema-example', () => {
     test('will fail when simple example is invalid', async () => {
       const results = await s.run({
         openapi: '3.0.2',
-        components: {
+        [field]: {
           schemas: {
             xoxo: {
               type: 'string',
@@ -57,14 +57,14 @@ describe('oas3-valid-schema-example', () => {
       ]);
     });
 
-    test('will pass for valid parents examples which contain invalid child examples', async () => {
+    test('will fail for valid parents examples which contain invalid child examples', async () => {
       const results = await s.run({
         openapi: '3.0.2',
         info: {
           version: '1.0.0',
           title: 'Swagger Petstore',
         },
-        components: {
+        [field]: {
           schemas: {
             post: {
               schema: {
@@ -108,7 +108,7 @@ describe('oas3-valid-schema-example', () => {
           code: 'oas3-valid-schema-example',
           message: '`example` property type should be string',
           path: [
-            'components',
+            field,
             'schemas',
             'post',
             'schema',
@@ -129,7 +129,7 @@ describe('oas3-valid-schema-example', () => {
     test('will pass when complex example is used ', async () => {
       const results = await s.run({
         openapi: '3.0.2',
-        components: {
+        [field]: {
           schemas: {
             xoxo: {
               type: 'object',
@@ -206,497 +206,9 @@ describe('oas3-valid-schema-example', () => {
     test('will error with totally invalid input', async () => {
       const results = await s.run({
         openapi: '3.0.2',
-        components: {
+        [field]: {
           schemas: {
             xoxo: {
-              type: 'object',
-              properties: {
-                url: {
-                  type: 'string',
-                },
-                width: {
-                  type: 'integer',
-                },
-                height: {
-                  type: 'integer',
-                },
-              },
-              required: ['url'],
-              example: {
-                url2: 'images/38.png',
-                width: 'coffee',
-                height: false,
-              },
-            },
-          },
-        },
-      });
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property should have required property `url`',
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-  });
-
-  describe('parameters', () => {
-    test('will pass when simple example is valid', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            schema: {
-              type: 'string',
-              example: 'doggie',
-            },
-          },
-        ],
-      });
-      expect(results).toHaveLength(0);
-    });
-
-    test('will fail for valid parents examples which contain invalid child examples', async () => {
-      const results = await s.run({
-        swagger: '2.0',
-        info: {
-          version: '1.0.0',
-          title: 'Swagger Petstore',
-        },
-        paths: {
-          '/pet': {
-            post: {
-              parameters: [
-                {
-                  in: 'body',
-                  name: 'body',
-                  required: true,
-                  schema: {
-                    type: 'object',
-                    example: {
-                      a: {
-                        b: {
-                          c: 'foo',
-                        },
-                      },
-                    },
-                    properties: {
-                      a: {
-                        type: 'object',
-                        example: {
-                          b: {
-                            c: 'foo',
-                          },
-                        },
-                        properties: {
-                          b: {
-                            type: 'object',
-                            properties: {
-                              c: {
-                                type: 'string',
-                                example: 12345,
-                              },
-                            },
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              ],
-              responses: {
-                '200': {
-                  description: 'OK',
-                },
-              },
-            },
-          },
-        },
-      });
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be string',
-          path: [
-            'paths',
-            '/pet',
-            'post',
-            'parameters',
-            '0',
-            'schema',
-            'properties',
-            'a',
-            'properties',
-            'b',
-            'properties',
-            'c',
-            'example',
-          ],
-          range: expect.any(Object),
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-
-    test('will not fail if an actual property is called example', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            in: 'body',
-            type: 'object',
-            properties: {
-              example: {
-                description: 'an actual field called example...',
-                type: 'string',
-              },
-            },
-            example: {
-              example: 'what is gonna happen',
-            },
-          },
-        ],
-      });
-
-      expect(results).toHaveLength(0);
-    });
-
-    test('will not fail if an actual property is called example and there is also type/format property', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            type: 'object',
-            properties: {
-              example: {
-                type: 'string',
-                example: 'abc',
-              },
-              type: {
-                type: 'number',
-                example: 123,
-              },
-              format: 'plain text',
-            },
-          },
-        ],
-      });
-
-      expect(results).toHaveLength(0);
-    });
-
-    test('will fail when simple example is invalid', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            schema: {
-              type: 'string',
-              example: 123,
-            },
-          },
-        ],
-      });
-      expect(results).toEqual([
-        expect.objectContaining({
-          severity: DiagnosticSeverity.Error,
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be string',
-        }),
-      ]);
-    });
-
-    test('will pass when complex example is used ', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            schema: {
-              type: 'object',
-              properties: {
-                url: {
-                  type: 'string',
-                },
-                width: {
-                  type: 'integer',
-                },
-                height: {
-                  type: 'integer',
-                },
-              },
-              required: ['url'],
-              example: {
-                url: 'images/38.png',
-                width: 100,
-                height: 100,
-              },
-            },
-          },
-        ],
-      });
-
-      expect(results).toHaveLength(0);
-    });
-
-    test('will fail when complex example is used', async () => {
-      const data = {
-        parameters: [
-          {
-            Heh: {
-              schema: {
-                type: 'number',
-                example: 4,
-              },
-            },
-            Abc: {
-              schema: {
-                type: 'object',
-                properties: {
-                  id: {
-                    type: 'integer',
-                    format: 'int64',
-                  },
-                  name: {
-                    type: 'string',
-                  },
-                  abc: {
-                    type: 'number',
-                    example: '5',
-                  },
-                },
-                required: ['name'],
-                example: {
-                  name: 'Puma',
-                  id: 1,
-                },
-              },
-            },
-          },
-        ],
-      };
-
-      const results = await s.run(data);
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be number',
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-
-    test('will error with totally invalid input', async () => {
-      const results = await s.run({
-        parameters: [
-          {
-            schema: {
-              type: 'object',
-              properties: {
-                url: {
-                  type: 'string',
-                },
-                width: {
-                  type: 'integer',
-                },
-                height: {
-                  type: 'integer',
-                },
-              },
-              required: ['url'],
-              example: {
-                url2: 'images/38.png',
-                width: 'coffee',
-                height: false,
-              },
-            },
-          },
-        ],
-      });
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property should have required property `url`',
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-  });
-
-  describe('headers', () => {
-    test('will pass when simple example is valid', async () => {
-      const results = await s.run({
-        openapi: '3.0.2',
-        headers: {
-          xoxo: {
-            schema: {
-              type: 'string',
-              example: 'doggie',
-            },
-          },
-        },
-      });
-      expect(results).toHaveLength(0);
-    });
-
-    test('will fail when simple example is invalid', async () => {
-      const results = await s.run({
-        openapi: '3.0.2',
-        headers: {
-          xoxo: {
-            schema: {
-              type: 'string',
-              example: 123,
-            },
-          },
-        },
-      });
-      expect(results).toEqual([
-        expect.objectContaining({
-          severity: DiagnosticSeverity.Error,
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be string',
-        }),
-      ]);
-    });
-
-    test('will fail for valid parents examples which contain invalid child examples', async () => {
-      const results = await s.run({
-        openapi: '3.0.2',
-        headers: {
-          xoxo: [
-            {
-              schema: {
-                type: 'object',
-                example: {
-                  a: {
-                    b: {
-                      c: 'foo',
-                    },
-                  },
-                },
-                properties: {
-                  a: {
-                    type: 'object',
-                    example: {
-                      b: {
-                        c: 'foo',
-                      },
-                    },
-                    properties: {
-                      b: {
-                        type: 'object',
-                        properties: {
-                          c: {
-                            type: 'string',
-                            example: 12345,
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          ],
-        },
-      });
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be string',
-          path: ['headers', 'xoxo', '0', 'schema', 'properties', 'a', 'properties', 'b', 'properties', 'c', 'example'],
-          range: expect.any(Object),
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-
-    test('will pass when complex example is used ', async () => {
-      const results = await s.run({
-        openapi: '3.0.2',
-        headers: {
-          xoxo: {
-            schema: {
-              type: 'object',
-              properties: {
-                url: {
-                  type: 'string',
-                },
-                width: {
-                  type: 'integer',
-                },
-                height: {
-                  type: 'integer',
-                },
-              },
-              required: ['url'],
-              example: {
-                url: 'images/38.png',
-                width: 100,
-                height: 100,
-              },
-            },
-          },
-        },
-      });
-
-      expect(results).toHaveLength(0);
-    });
-
-    test('will fail when complex example is used ', async () => {
-      const data = {
-        openapi: '3.0.2',
-        headers: {
-          Heh: {
-            schema: {
-              type: 'number',
-              example: 4,
-            },
-          },
-          Abc: {
-            schema: {
-              type: 'object',
-              properties: {
-                id: {
-                  type: 'integer',
-                  format: 'int64',
-                },
-                name: {
-                  type: 'string',
-                },
-                abc: {
-                  type: 'number',
-                  example: '5',
-                },
-              },
-              required: ['name'],
-              example: {
-                name: 'Puma',
-                id: 1,
-              },
-            },
-          },
-        },
-      };
-
-      const results = await s.run(data);
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          code: 'oas3-valid-schema-example',
-          message: '`example` property type should be number',
-          severity: DiagnosticSeverity.Error,
-        }),
-      ]);
-    });
-
-    test('will error with totally invalid input', async () => {
-      const results = await s.run({
-        openapi: '3.0.2',
-        headers: {
-          xoxo: {
-            schema: {
               type: 'object',
               properties: {
                 url: {
@@ -732,7 +244,7 @@ describe('oas3-valid-schema-example', () => {
     test('does not report example mismatches for unknown AJV formats', async () => {
       const results = await s.run({
         openapi: '3.0.2',
-        headers: {
+        [field]: {
           xoxo: {
             schema: {
               type: 'object',
@@ -750,7 +262,9 @@ describe('oas3-valid-schema-example', () => {
 
       expect(results).toEqual([]);
     });
+  });
 
+  describe('headers', () => {
     test.each([
       ['byte', '1'],
       ['int32', 2 ** 31],
